Added ColorMapper class to support working with colors. Added palette_select option type for tiles. Added ability to set background color of a cell in the main table. The CentroidCluster tile now colors the background of cells along with the codes. Added ClusterViewer tile.
Updated build_html_table_from_data_list to allow for either clicking of rows or cells.
Also added styles for :hover and :active so that tables are more interactive.

Dealt with the problem that sometimes saved tiles wouldn't appear when a saved project was loaded.
There were a couple of problems. But the main one was that the tiles were loaded after a call to set_visible_doc.
The call to set_visible_doc didn't always complete in time.

Added some extra error handling here and there.