This update is about allowing multiple selections in the library manager
Edit tags for multiple rows
Drag and drop to and from selection of multiple rows.
Some buttons operate on multiple rows

Combining multiple collections when selecting them creates an
entirely new collection.

Added ability to compare tiles in the library

Some refactoring of handling of datestrings

When appending a document to a collection now ensures name is unique